build: restructure for cmake once more
