--- libseam_carving/coord.cpp
@@ -1,22 +1,0 @@
-//
-// Created by Andy Mina on 3/7/23.
-//
-
-#include <seam_carving/coord.hpp>
-
-namespace seam_carving {
-    Coord::Coord(const int& row, const int& col):
-            row_{row}, col_{col}
-    { }
-
-    void to_json(nlohmann::json& j, const Coord& coord) {
-        j = nlohmann::json{
-                {"row", coord.row()},
-                {"col", coord.col()}
-        };
-    }
-
-    void from_json(const nlohmann::json& j, Coord& coord) {
-        coord = Coord(j.at("row"), j.at("col"));
-    }
-}
