+++ libseam_carving/seam_carving/energy.hpp
@@ -37,7 +37,7 @@ namespace seam_carving::energy {
      * the brightness is proportional to the energy of the pixel. Seams can be
      * created by following the path of least energy. The vertical path of least
      * energy is found by the following algorithm:
-     *    1. initalize the bottom row with weights equal to their energy
+     *    1. initialize the bottom row with weights equal to their energy
      *    2. move to the next row above (row - 1 -> row)
      *    3. loop through each pixel in the next row
      *    4. compute the weights adjacent pixels in the row underneath, specified
