+++ libseam_carving/seam_carving/tests/energy_data.hpp
@@ -6,7 +6,6 @@
 #define SEAM_CARVING_ENERGY_DATA_HPP
 
 // project
-#include <seam_carving/tests/matrix_data.hpp>
 #include <seam_carving/tests/test_data.hpp>
 
 // 3rd party
@@ -14,17 +13,12 @@
 #include <opencv2/core.hpp>
 #include <gtest/gtest.h>
 
-// std
-#include <fstream>
-#include <sstream>
-
 namespace seam_carving::tests {
-    struct EnergyData : TestData {
+    struct EnergyData : public TestData {
         cv::Mat sobel_matrix;
         cv::Mat vertical_map_matrix, horizontal_map_matrix;
     };
 
-    void from_json(const nlohmann::json& j, EnergyData& energy_data);
     class EnergyTest : public testing::TestWithParam<EnergyData> { };
 }
